csr: Check privilege level and read-only bits from csr address in unit
